requirements.txt
openpyxl
pytest
pytest-django

